Update ImageProcessingTensorFlowBirdClassification.py
This was an example dataset that we adjusted parameters and did testing to understand how to handle image processing with TensorFlow in python. The program creates a model to process images to decide if the image is of a bird or if the image is not of a bird. The project demonstrates using a Neural Network to classify images.